- rearranged detected test files and modules to avoid pytest   load dependency errors;
